Summary: Extension to SCEVExpander to handle HIR blob types Reviewed by: Chawla, Pankaj; Sharma, Kamal G Rurl: https://ir-codecollab.intel.com/ui#review:id=25890 Old/split llvm commit ID: 936ede58ed41dc77f8a79b3087f1e7e450464174

+++ llvm/include/llvm/IR/Intel_LoopIR/CanonExpr.h
@@ -438,7 +438,7 @@ public:
   int64_t getBlobCoeff(const_blob_iterator CBlobI) const;
 
   /// \brief Returns the blob index of the only blob.
-  int64_t getSingleBlobIndex() const {
+  unsigned getSingleBlobIndex() const {
     assert((numBlobs() == 1) && "Canon expr does not contain single blob!");
     return BlobCoeffs[0].Index;
   }
